requirements.txt
accelerate==0.22.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
blis==0.7.10
catalogue==2.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.4.1
confection==0.1.3
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
Cython==3.0.5
datasets==2.16.1
dill==0.3.7
en-core-web-sm==3.6.0
filelock==3.12.3
Flask==2.3.3
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.0
huggingface-hub==0.20.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
langcodes==3.3.0
lit==16.0.6
MarkupSafe==2.1.3
matplotlib==3.7.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.9
networkx==3.1
nltk==3.8.1
numpy==1.24.4
openai==0.28.0
packaging==23.1
pandas==2.0.3
pathy==0.10.2
Pillow==10.0.0
pip==23.2.1
preshed==3.0.8
psutil==5.9.5
pyarrow==15.0.0
pyarrow-hotfix==0.6
pydantic==2.3.0
pydantic_core==2.6.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==68.0.0
six==1.16.0
smart-open==6.4.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.7
sympy==1.12
thinc==8.1.12
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.31.0
triton==2.0.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
Unidecode==1.3.6
urllib3==2.0.4
wasabi==1.1.2
Werkzeug==2.3.7
wheel==0.38.4
xformers==0.0.23.post1
xxhash==3.4.1
yarl==1.9.2
zipp==3.16.2
